refactor(project-structure): reorganize project structure by moving files to improve maintainability
