+++ src/app/(home)/components/Hero/HeroSection.tsx
