+++ src/app/(home)/components/Projects/index.ts
@@ -1,0 +1,1 @@
+export * from './ProjectsSection';
