+++ src/app/(home)/components/Stack/StackSection.tsx
