+++ src/app/projects/[slug]/(components)/index.ts
@@ -1,0 +1,1 @@
+export * from './NextProjectBtn';
