+++ src/components/common/index.ts
@@ -1,1 +1,4 @@
+export * from './LogoLink';
+export * from './NotionRichText';
 export * from './Socials';
+export * from './ThemedImg';
