+++ src/components/index.ts
@@ -1,4 +1,3 @@
 export * from './common';
 export * from './layout';
-export * from './sections';
 export * from './ui';
