+++ src/components/layout/Header/components/index.ts
@@ -1,0 +1,3 @@
+export * from './NavLinks';
+export * from './ResumeBtn';
+export * from './ThemeSwitcher';
