+++ src/components/layout/index.ts
@@ -1,3 +1,3 @@
 export * from './Cta';
 export * from './Footer';
-export * from './Header';
+export * from './Header/Header';
