+++ src/components/ui/index.ts
@@ -1,11 +1,3 @@
 export * from './Button';
-export * from './Card';
 export * from './Container';
-export * from './LogoLink';
-export * from './NavLinks';
-export * from './NextProjectBtn';
-export * from './NotionRichText';
-export * from './ResumeBtn';
-export * from './ThemedImg';
-export * from './ThemeSwitcher';
 export * from './Typography';
